sim: split stepper input-discontinuity tests into submodule
Move the three SimStepper input-discontinuity tests out of
with_diffsol/tests/mod.rs into a new stepper_input_discontinuity
submodule, bringing mod.rs back under the 2000-line lint ceiling.
